+++ src/lib/components/InputWithIcon.tsx
@@ -2,6 +2,8 @@ import type { InputProps } from "@chakra-ui/react";
 import { Input, InputGroup, InputRightElement } from "@chakra-ui/react";
 import type { ChangeEvent } from "react";
 
+import { AmpEvent, AmpTrack } from "lib/services/amplitude";
+
 import { CustomIcon } from "./icon";
 
 interface InputWithIconProps {
@@ -9,12 +11,14 @@ interface InputWithIconProps {
   value: string;
   onChange: (e: ChangeEvent<HTMLInputElement>) => void;
   size?: InputProps["size"];
+  action?: string;
 }
 
 const InputWithIcon = ({
   placeholder,
   value,
   size,
+  action,
   onChange,
 }: InputWithIconProps) => (
   <InputGroup>
@@ -23,6 +27,7 @@ const InputWithIcon = ({
       value={value}
       onChange={onChange}
       size={size}
+      onClick={action ? () => AmpTrack(AmpEvent.USE_SEARCH_INPUT) : undefined}
     />
     <InputRightElement h="56px" alignItems="center" mr="1">
       <CustomIcon name="search" />